Move the whole thumbnail reading/writing machinery into the ImageIO module to be actually able to do type dispatch.
