+++ rtengine/image16.h
@@ -97,6 +97,8 @@ public:
     }
 
     void                 ExecCMSTransform(cmsHTRANSFORM hTransform);
+
+    bool writeThumbnail (const Glib::ustring& fname) override;
 };
 
 }
